# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Catalog Service API
  description: |
    Manages the RUBY MUSIC catalog: genres, stations, artists, albums and songs.
    Stations and Genres are distinct concepts: a Genre is a musical tag on songs;
    a Station is a themed listening room that aggregates songs from one or more genres.
    All write operations require admin privileges (enforced at gateway level).
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Local development
  - url: http://localhost:8080/api/v1/catalog
    description: Via API Gateway
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/catalog/stations:
    get:
      tags: [Stations]
      operationId: listStations
      summary: List stations paginated (admin)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of stations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationPage'
    post:
      tags: [Stations]
      operationId: createStation
      summary: Create a station (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StationRequest'
      responses:
        '201':
          description: Station created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationResponse'
        '409':
          description: Station name already exists
  /api/v1/catalog/stations/search:
    get:
      tags: [Stations]
      operationId: searchStations
      summary: Search active stations by name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationPage'
  /api/v1/catalog/stations/{id}:
    get:
      tags: [Stations]
      operationId: getStationById
      summary: Get station by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Station found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationResponse'
        '404':
          description: Not found
    put:
      tags: [Stations]
      operationId: updateStation
      summary: Update station (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StationRequest'
      responses:
        '200':
          description: Station updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationResponse'
    delete:
      tags: [Stations]
      operationId: deleteStation
      summary: Delete station (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Deleted
  /api/v1/catalog/stations/{id}/songs:
    get:
      tags: [Stations, Songs]
      operationId: getSongsByStation
      summary: Songs directly assigned to a station (paginated)
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPage'
        '404':
          description: Station not found

  # ── Artists ──────────────────────────────────────────────────────────────────
...
  /api/v1/catalog/songs/recommendations:
    get:
      tags: [Songs]
      operationId: getRecommendations
      summary: Get song recommendations based on a set of reference songs (for "Canciones recomendadas" in playlist detail)
      parameters:
        - name: songIds
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Recommended songs by genre overlap, excluding the input songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPage'
  /api/v1/catalog/songs/search:
    get:
      tags: [Songs]
      operationId: searchSongs
      summary: Search songs by title or artist name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPage'
  /api/v1/catalog/songs/{id}:
    get:
      tags: [Songs]
      operationId: getSongById
      summary: Get song by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Song found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongResponse'
        '404':
          description: Not found
    put:
      tags: [Songs]
      operationId: updateSong
      summary: Update song (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SongUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongResponse'
    delete:
      tags: [Songs]
      operationId: deleteSong
      summary: Delete song (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Deleted
  /api/v1/catalog/songs:
    get:
      tags: [Songs]
      operationId: listSongs
      summary: List all songs paginated (admin Canciones panel)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPage'
    post:
      tags: [Songs]
      operationId: createSong
      summary: Create song (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SongCreateRequest'
      responses:
        '201':
          description: Song created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
  schemas:
...
    GenreResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        gradientStart:
          type: string
        gradientEnd:
          type: string
        songCount:
          type: integer
          description: Number of songs tagged with this genre (computed)
        createdAt:
          type: string
          format: date-time
    StationRequest:
      type: object
      required: [name, genreId, gradientStart, gradientEnd, songIds]
      properties:
        name:
          type: string
          maxLength: 100
        genreId:
          type: string
          format: uuid
          description: Single genre that categorizes this station (shown in detail view)
        gradientStart:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        gradientEnd:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        isActive:
          type: boolean
        songIds:
          type: array
          minItems: 3
          items:
            type: string
            format: uuid
          description: Songs directly assigned to this station (minimum 3 required)
    StationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        genreId:
          type: string
          format: uuid
        genreName:
          type: string
        gradientStart:
          type: string
        gradientEnd:
          type: string
        songCount:
          type: integer
          description: Number of songs directly assigned to this station
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StationPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/StationResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer
...
    ArtistResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        photoUrl:
          type: string
        bio:
          type: string
        monthlyListeners:
          type: integer
          format: int64
        followersCount:
          type: integer
          format: int64
        isTop:
          type: boolean
        createdAt:
          type: string
          format: date-time
          description: Registration date shown in admin artist detail
...
    AlbumResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          $ref: '#/components/schemas/ArtistResponse'
        coverUrl:
          type: string
        releaseDate:
          type: string
          format: date
        totalStreams:
          type: integer
          format: int64
        songCount:
          type: integer
          description: Number of songs in this album (computed)
        stationId:
          type: string
          format: uuid
          description: Station this album belongs to (null if not assigned)
        stationName:
          type: string
          description: Display name of the station (null if not assigned)
...
    SongCreateRequest:
      type: object
      required: [title, artistId, coverUrl, audioUrl, duration]
      properties:
        title:
          type: string
        artistId:
          type: string
          format: uuid
        albumId:
          type: string
          format: uuid
        coverUrl:
          type: string
        audioUrl:
          type: string
        duration:
          type: integer
          description: Duration in seconds
        lyrics:
          type: string
        genreIds:
          type: array
          items:
            type: string
            format: uuid
    SongUpdateRequest:
      type: object
      properties:
        title:
          type: string
        coverUrl:
          type: string
        audioUrl:
          type: string
        lyrics:
          type: string
        genreIds:
          type: array
          items:
            type: string
            format: uuid
    SongResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          $ref: '#/components/schemas/ArtistResponse'
        album:
          $ref: '#/components/schemas/AlbumResponse'
        coverUrl:
          type: string
        audioUrl:
          type: string
        duration:
          type: integer
        lyrics:
          type: string
        playCount:
          type: integer
          format: int64
        likesCount:
          type: integer
          format: int64
        genres:
          type: array
          items:
            $ref: '#/components/schemas/GenreResponse'
    SongPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SongResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer
...